+++ src/main.js
@@ -186,8 +186,6 @@ import { loadState } from '@nextcloud/initial-state'
 			|| (!OCA.Onlyoffice.setting.sameTab && !isDefault)) {
 			OCA.Onlyoffice.SetDefaultUrl()
 			winEditor = window.open(url, '_blank')
-		} else if (isPublicShare() && OCA.Onlyoffice.isViewIsFile()) {
-			location.href = url
 		} else {
 			if (OCA.Onlyoffice.setting.enableSharing
 				&& !isPublicShare()
@@ -915,23 +913,26 @@ import { loadState } from '@nextcloud/initial-state'
 				return
 			}
 
-			const editorUrl = OC.generateUrl('apps/' + OCA.Onlyoffice.AppName + '/s/' + encodeURIComponent(getSharingToken()))
+			registerFileAction(new FileAction({
+				id: 'onlyoffice-public-open',
+				displayName: () => t(OCA.Onlyoffice.AppName, 'Open in ONLYOFFICE'),
+				iconSvgInline: () => AppDarkSvg,
+				enabled: (files) => {
+					if (Permission.READ !== (files[0].permissions & Permission.READ)) { return false }
 
-			if (_oc_appswebroots.richdocuments
-				|| (_oc_appswebroots.files_pdfviewer && extension === 'pdf')
-				|| (_oc_appswebroots.text && extension === 'txt')) {
+					return true
+				},
+				exec(file, view, dir) {
+					OCA.Onlyoffice.FileClickExec(file, view, dir, false)
+				},
+			}))
 
-				const button = document.createElement('a')
-				button.href = editorUrl
-				button.className = 'onlyoffice-public-open button'
-				button.innerText = t(OCA.Onlyoffice.AppName, 'Open in ONLYOFFICE')
+			if (config.def
+				&& !_oc_appswebroots.richdocuments
+				&& !(_oc_appswebroots.files_pdfviewer && extension === 'pdf')
+				&& !(_oc_appswebroots.text && extension === 'txt')) {
+				const editorUrl = OC.generateUrl('apps/' + OCA.Onlyoffice.AppName + '/s/' + encodeURIComponent(getSharingToken()))
 
-				if (!OCA.Onlyoffice.setting.sameTab) {
-					button.target = '_blank'
-				}
-
-				$('#preview').prepend(button)
-			} else {
 				OCA.Onlyoffice.frameSelector = '#onlyofficeFrame'
 				const container = document.createElement('div')
 				container.classList.add('onlyoffice-iframe-container')